Some major refactor

+++ Application/Features/Commands/Projects/DeleteProject/DeleteProjectCommandHandler.cs
@@ -1,4 +1,5 @@
-﻿using Domain.Interfaces;
+﻿using Application.Interfaces;
+using Domain.Interfaces;
 using MediatR;
 
 namespace Application.Features.Commands.Projects.DeleteProject
@@ -6,10 +7,12 @@ namespace Application.Features.Commands.Projects.DeleteProject
     public class DeleteProjectCommandHandler : IRequestHandler<DeleteProjectCommand, bool>
     {
         private readonly IProjectRepository _projectRepository;
+        private readonly IProjectAuthorizationService _authorizationService;
 
-        public DeleteProjectCommandHandler(IProjectRepository projectRepository)
+        public DeleteProjectCommandHandler(IProjectRepository projectRepository, IProjectAuthorizationService authorizationService)
         {
             _projectRepository = projectRepository;
+            _authorizationService = authorizationService;
         }
 
         public async Task<bool> Handle(DeleteProjectCommand request, CancellationToken cancellationToken)
@@ -20,6 +23,14 @@ namespace Application.Features.Commands.Projects.DeleteProject
             {
                 return false;
             }
+
+            // Authorization check
+            var isOwner = await _authorizationService.IsUserOwnerAsync(request.Id, request.UserId);
+            if (!isOwner)
+            {
+                return false;
+            }
+
             // Delete the project
             var deleted = await _projectRepository.DeleteProjectAsync(request.Id);
             return deleted;
